# [Profiler] Prefer TSC to wall clock when available